Commands Registry created
Added a commands class that is intended to validate functions/callables
that want to be used in procedures. It inspects the function definition
to make sure that it is of the right format.

Todo - add in some convenience functions, or maybe caching, of the
parameters that it requires for handing to the GUI or whatnot.