Instances can retrieve variables and Procedure prep
Instances are in their own python file sicne it was starting to get a
bit long.
Instances can look up the value of a workvariable given its name,
including grabbing it from the workflow.

procedures has been rename to procedure_runner, and the ProcedureStep
was moved into the workflows files.
 - The step is such a tiny container, that I do not want to deal with
   more circular dependencies for the runner