Cleanup PipelineManager and fix loop bug
Cleaned up the logic to be nicer looking and pythonic for the
PipelineManager when fixing a bug I realized.
The global_instances array cannot be directly iterated upon while
processing commans because they will add and remove from that
array, and thus we must make a copy of what we want for iteration.