Commands in procedures can specify unions of types
Added in type support for unions of types (String|Integer) and the
procedure runner will attempt to dereference and coerce just like the
previous non union type annotations. This did require some changes to
the Commands register function, but should be a transparent change.

This also got me to actually clean up the ProcedureRunner to not just
have a giant ugly single method to run a single step.
The variable logic for coercion and dereferencing was made much cleaner
and the error printing was made more consistent.

Also! I found a couple extra places in the ProcedureRunner that were not
handling potential sources of exceptiosn being thrown. That should be
bullet proof now in terms of catching everything and not taking down the
server/daemon.

Also converted the CommandReturnStatus into a Flag instead of an Enum.
I wanted to have extra information to help the jump/goto commands tell
the ProcedureRunner to not bother autoincrementing the current step
since it would have just been changed to something correct.
This will also allow commands such as wait_for_download_to_finish() to
exist in a simpler form rather than constant jumps and variable tracking
of state.