Full test of a non-trivial workflow instance
Some minor bugs, some more commands, but the true value is the main file
with a bunch of manually entered commands that loop and increment and
stuff. The pipeline manager and procedure runner seem to be working as
expected now, even some accidental (not commit) tests of the workflow
erroring and the console providing useful information to debug.

Some other minor changes:
- jump_to_procedure bug fix to use the .value for the name
- some extra minor commands and some math commadns
- Instance variable get also allows VariableName for usage simplicity
- bugfix of procedure runner handling return status codes mixed with
  other values
- Workflow __repr__() now prints some procedures debug info