proof of concept of the UI addon
This is a great milestone of having a server working with addons
dynamically registering endpoints to serve up the static content files
that are in the module.

- serves index.html as a static path and static filepath
- serves javascript files from a static location with proper path
  normalization and checking to ensure no simple .. escapement
- Basic webpage with a quick html template to translate the json into
  something a bit more human readable.

It is unclear if i want to continue down the html template path that i
have started in the html file, but that is totally fine! I can play with
the front end as much as I want, and it is all contained to this addon
to ensure I don't accidentally mess up core code, and that I never
introduce dependencies to the core from the front end.