UI Global Variables Section + Editing
Added functionality to add and edit new global variables.
It is just inline editing since they are such simple items.

Also bugfix for the globals get api so it properly calls the
json_savable() on the variable

Homoginized the global variables api endpoints to all be
global_variables instead of global_vars